Fixed single file encoding and decoding.
Downloaded the test files from the yenc.org website and tweaked the code
so that it conforms. For encoding, the original file is a message with
various internet headers (e.g. email or usenet) so the verification
required a copy of the file that contains only from the =ybegin to the
=yend tags.